+++ R/conversion.R
@@ -11,15 +11,14 @@
 #'     \item \code{make_units}, constructs units from bare expressions.
 #'     \code{make_units(m/s)} is equivalent to \code{as_units(quote(m/s))}.
 #'     \item \code{set_units}, a pipe-friendly version of \code{`units<-`}. By
-#'     default it operates with bare expressions like \code{make_unit}, but this
+#'     default it operates with bare expressions, but this
 #'     behavior can be disabled by a specifying \code{mode = "standard"} or setting
 #'     \code{units_options(set_units_mode = "standard")}.
 #' }
 #'
 #' @param x numeric vector, or object of class \code{units}.
 #' @param value object of class \code{units} or \code{symbolic_units}, or in the
-#' case of \code{set_units} expression with symbols that can be resolved in
-#' \link{ud_units} (see examples).
+#' case of \code{set_units} expression with symbols (see examples).
 #'
 #' @return An object of class \code{units}.
 #'
@@ -40,10 +39,10 @@
 `units<-.numeric` <- function(x, value) {
   if(is.null(value))
     return(x)
- 
+
   if(!inherits(value, "units") && !inherits(value, "symbolic_units"))
     value <- as_units(value)
-  
+
   if (inherits(value, "units")) {
 	if (any(is.na(value)))
 	  stop("a missing value for units is not allowed")
@@ -53,7 +52,7 @@
 	  warning(paste("numeric value", unclass(value), "is ignored in unit assignment"))
     value <- units(value)
   }
- 
+
   attr(x, "units") = value
   class(x) <- "units"
   x
@@ -61,35 +60,35 @@
 
 #' @name units
 #' @export
-#' 
+#'
 #' @examples
 #' a <- set_units(1:3, m/s)
-#' units(a) <- with(ud_units, km/h)
+#' units(a) <- make_units(km/h)
 #' a
 #' # convert to a mixed_units object:
 #' units(a) = c("m/s", "km/h", "km/h")
 #' a
 `units<-.units` <- function(x, value) {
-  
+
   if(is.null(value))
     return(drop_units(x))
-  
+
   if(!inherits(value, "units") && !inherits(value, "symbolic_units")) {
 	if ((is.character(value) && length(value) > 1))
 	  return(set_units(mixed_units(x), value))
     value <- as_units(value)
   }
-  
+
   dimx = dim(x)
   if (inherits(value, "units")) {
     if (!identical(as.numeric(value), 1))
       x <- .as.units(unclass(x) * unclass(value), units(x))
     value <- units(value)
   }
-  
+
   if (identical(units(x), value)) # do nothing; possibly user-defined units:
     return(x)
-  
+
   str1 <- as.character(units(x))
   str2 <- as.character(value)
 
@@ -108,9 +107,9 @@
 #' @name units
 #' @export
 `units<-.logical` <- function(x, value) {
-  if (!all(is.na(x))) 
+  if (!all(is.na(x)))
     stop("x must be numeric, non-NA logical not supported")
-  
+
   x <- as.numeric(x)
   units(x) <- value
   x
@@ -143,7 +142,7 @@ as.data.frame.units <- function(x, row.names = NULL, optional = FALSE, ...) {
 
 #' @export
 as.list.units <- function(x, ...)
-  mapply(set_units, unclass(x), x, mode="standard", SIMPLIFY=FALSE)
+  lapply(NextMethod(), set_units, units(x), mode="standard")
 
 #' convert units object into difftime object
 #'
@@ -151,9 +150,9 @@ as.list.units <- function(x, ...)
 #'
 #' @export
 #' @examples
-#' 
-#' t1 = Sys.time() 
-#' t2 = t1 + 3600 
+#'
+#' t1 = Sys.time()
+#' t2 = t1 + 3600
 #' d = t2 - t1
 #' du <- as_units(d)
 #' dt = as_difftime(du)
@@ -182,10 +181,10 @@ as_difftime <- function(x) {
 # #' @return object of class hms
 # #' @examples
 # #' if (require(hms)) {
-# #'  as.hms(1:10 * with(ud_units, s))
-# #'  as.hms(1:10 * with(ud_units, min))
-# #'  as.hms(1:10 * with(ud_units, h))
-# #'  as.hms(1:10 * with(ud_units, d))
+# #'  as.hms(1:10 * make_units(s))
+# #'  as.hms(1:10 * make_units(min))
+# #'  as.hms(1:10 * make_units(h))
+# #'  as.hms(1:10 * make_units(d))
 # #' }
 # #' @export
 # as.hms.units = function(x, ...) {