+++ demo/year.R
@@ -10,8 +10,8 @@ NISTyearSiderealTOsec(1) / (24 * 3600) # number of mean solar days days in 1 sid
 
 # ... and in udunits2:
 library(units)
-as.units(make_unit("common_year"), with(ud_units, d))
-as.units(make_unit("leap_year"), with(ud_units, d)) 
-as.units(make_unit("Gregorian_year"), with(ud_units, d))
-as.units(make_unit("Julian_year"), with(ud_units, d))
-as.units(with(ud_units, yr), with(ud_units, d)) # yr = tropical:
+as_units("common_year")
+as_units("leap_year")
+as_units("Gregorian_year")
+as_units("Julian_year")
+as_units("yr") # yr = tropical: